asr_diarization: Modified reverberation script by moving some functions to library and adding extra options
